refactor project to use modules
