+++ index.html
@@ -58,6 +58,6 @@
       <p>&copy; Copyright 2022 All rights reserved</p>
     </footer>
   </main> 
-    <script src="index.js"></script>
+    <script type="module" src="./main/index.js"></script>
 </body>
 </html>
