+++ modules/books.js
@@ -1,0 +1,46 @@
+let books = JSON.parse(localStorage.getItem('books'));
+// add class
+export default class Book {
+  constructor(title, author, id) {
+    this.title = title;
+    this.author = author;
+    this.id = id;
+  }
+
+  newBook() {
+    const { title, author, id } = this;
+    const bookList = { title, author, id };
+    books = JSON.parse(localStorage.getItem('books'));
+    if (title === '' || author === '') {
+      document.getElementById('error').innerHTML = 'Required';
+    } else if (books !== null) {
+      books.push(bookList);
+      localStorage.setItem('books', JSON.stringify(books));
+      books = JSON.parse(localStorage.getItem('books'));
+    } else {
+      books = [];
+      books.push(bookList);
+      localStorage.setItem('books', JSON.stringify(books));
+      books = JSON.parse(localStorage.getItem('books'));
+    }
+  }
+
+  // remove
+  remove() {
+    const { id } = this;
+    books = books.filter((book) => {
+      if (book.id === id) {
+        return false;
+      }
+      return true;
+    });
+
+    localStorage.setItem('books', JSON.stringify(books));
+  }
+}
+
+if (books !== null) {
+  books.forEach((book) => {
+    displayBooks(book.title, book.author, book.id);
+  });
+}
